% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppaddColumnOfOnes.R
\name{RcppaddColumnOfOnes}
\alias{RcppaddColumnOfOnes}
\title{This function add one column of ones to input matrix mat of a linear regression model.}
\usage{
RcppaddColumnOfOnes(mat)
}
\arguments{
\item{mat}{Numeric matrix of independent variables.}
}
\value{
Numeric matrix of standard format of linear regression.
}
\description{
This function add one column of ones to input matrix mat of a linear regression model.
}
