% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppgetBetaHat.R
\name{RcppgetBetaHat}
\alias{RcppgetBetaHat}
\title{This function estimates the coefficients (Beta Hat) of a linear regression model.}
\usage{
RcppgetBetaHat(mat, y)
}
\arguments{
\item{mat}{Numeric matrix of independent variables.}

\item{y}{Numeric vector of dependent variable.}
}
\value{
Numeric vector representing the estimated coefficients.
}
\description{
This function estimates the coefficients (Beta Hat) of a linear regression model.
}
