% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppgetF.R
\name{RcppgetF}
\alias{RcppgetF}
\title{This function calculates the F-statistic and p-value for a F-test of a linear regression model.}
\usage{
RcppgetF(mat, y)
}
\arguments{
\item{mat}{Numeric matrix of independent variables.}

\item{y}{Numeric vector of dependent variable.}
}
\value{
A list containing results of F-test of the model.
}
\description{
This function calculates the F-statistic and p-value for a F-test of a linear regression model.
}
