% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppgetResidual.R
\name{RcppgetResidual}
\alias{RcppgetResidual}
\title{This function calculates the residuals(Y-Yhat) of a linear regression model.}
\usage{
RcppgetResidual(mat, y)
}
\arguments{
\item{mat}{Numeric matrix of independent variables.}

\item{y}{Numeric vector of dependent variable.}
}
\value{
Numeric vector representing the residuals.
}
\description{
This function calculates the residuals(Y-Yhat) of a linear regression model.
}
