% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppgetRsquare.R
\name{RcppgetRsquare}
\alias{RcppgetRsquare}
\title{This function calculates the Rsquare (coefficient of determination) of a linear regression model.}
\usage{
RcppgetRsquare(mat, y)
}
\arguments{
\item{mat}{Numeric matrix of independent variables.}

\item{y}{Numeric vector of dependent variable.}
}
\value{
A scalar representing the R^2 of the model.
}
\description{
This function calculates the Rsquare (coefficient of determination) of a linear regression model.
}
