% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppgetT.R
\name{RcppgetT}
\alias{RcppgetT}
\title{This function calculates the T-statistic, standard error, p-value and statistical decisions for a T-test of a linear regression model.}
\usage{
RcppgetT(mat, y)
}
\arguments{
\item{mat}{Numeric matrix of independent variables.}

\item{y}{Numeric vector of dependent variable.}
}
\value{
A list containing results of T-test of the model.
}
\description{
This function calculates the T-statistic, standard error, p-value and statistical decisions for a T-test of a linear regression model.
}
